drivers_app/seeker/__init__.py

from sqlalchemy import create_engine
from sqlalchemy.orm import sessionmaker

from config import config


mysql_conn = config['dev'].SQLALCHEMY_DATABASE_URI
engine = create_engine(mysql_conn)
DBSession = sessionmaker(bind=engine)
session = DBSession()


from .core import run_carbens_spider
